requirements.txt
ai21==2.2.0
ai21-tokenizer==0.9.0
aiohttp==3.9.3
aiosignal==1.3.1
aiosmtpd==1.4.4.post2
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.3
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
atpublic==4.0
attrs==23.2.0
Authlib==1.2.1
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.15.0
azure-mgmt-core==1.4.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.0.1
azure-mgmt-subscription==3.1.1
Babel==2.14.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.7.0
boto3==1.33.13
botocore==1.33.13
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
cryptography==42.0.5
dataclasses-json==0.6.4
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.5.0
email-validator==2.1.0.post1
eventlet==0.34.2
executing==2.0.1
filelock==3.13.4
Flask==2.3.3
flask-babel==4.0.0
Flask-Compress==1.14
Flask-Gravatar==0.5.0
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.3.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.3.1
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpagentparser==1.9.5
httpcore==1.0.4
httpie==3.2.2
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
importlib_resources==6.4.0
ipykernel==6.29.0
ipython==8.21.0
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.3.1
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonformatter==0.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
keyring==24.3.1
langchain==0.1.16
langchain-ai21==0.1.3
langchain-community==0.0.32
langchain-core==0.1.42
langchain-experimental==0.0.57
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.23
ldap3==2.9.1
mailchimp_marketing @ git+https://github.com/mailchimp/mailchimp-marketing-python.git@3305fa45b3f436767a539c5fba9cb2b0a083d761
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
migrations==0.1.1
more-itertools==10.2.0
msal==1.28.0
msal-extensions==1.1.0
msrest==0.7.1
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.13.3
orjson==3.9.15
packaging==23.2
paramiko==3.4.0
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
pgvector==0.2.5
pillow==10.3.0
platformdirs==4.2.0
portalocker==2.8.2
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
PyJWT==2.8.0
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.1.2
pypdf==4.1.0
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2023.4
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
redis==5.0.1
Redis-Sentinel-Url==1.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rq==1.15.1
rq-dashboard==0.6.7.2
rsa==4.9
s3transfer==0.8.2
SecretStorage==3.3.3
sentencepiece==0.2.0
setuptools==69.2.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
speaklater3==1.4
SQLAlchemy==2.0.25
sqlparse==0.4.4
sshtunnel==0.4.0
stack-data==0.6.3
tenacity==8.2.3
tesserocr==2.6.3
tiktoken==0.6.0
tokenizers==0.15.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typer==0.9.4
types-python-dateutil==2.8.19.20240106
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.9.0
ua-parser==0.18.0
uritemplate==4.1.1
urllib3
user-agents==2.2.0
wcwidth==0.2.13
Werkzeug==2.3.8
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.4
